before day 8 (refactoring some methods & classes; managing some user-database issues)

+++ src/Controller/LicensePlateController.php
@@ -6,11 +6,13 @@ use App\Entity\LicensePlate;
 use App\Form\LicensePlateType;
 use App\Repository\LicensePlateRepository;
 use App\Service\ActivityService;
+use App\Service\LicensePlateService;
 use App\Service\MailerService;
 use Doctrine\ORM\NonUniqueResultException;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Mailer\Exception\TransportExceptionInterface;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\String\UnicodeString;
 
@@ -26,19 +28,17 @@ class LicensePlateController extends AbstractController
     }
 
     /**
-     * @throws NonUniqueResultException
+     * @throws NonUniqueResultException|TransportExceptionInterface
      */
     #[Route('/new', name: 'license_plate_new', methods: ['GET', 'POST'])]
-    public function new(Request $request, ActivityService $activity, MailerService $mailer, LicensePlateRepository $licensePlateRepository): Response
+    public function new(Request $request, ActivityService $activity, MailerService $mailer, LicensePlateRepository $licensePlateRepository, LicensePlateService $licensePlateService): Response
     {
         $licensePlate = new LicensePlate();
         $form = $this->createForm(LicensePlateType::class, $licensePlate);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $licensePlate->setLicensePlate((new UnicodeString($licensePlate->getLicensePlate()))->camel()->upper());
-
-            $this->addFlash('success', 'Your add a new car!');
+            $licensePlate->setLicensePlate($licensePlateService->normalizeLicensePlate($licensePlate->getLicensePlate()));
 
             $entry = $licensePlateRepository->findOneBy(['license_plate' => $licensePlate->getLicensePlate()]);
 
@@ -53,28 +53,41 @@ class LicensePlateController extends AbstractController
                 $blocker = $activity->whoBlockedMe($licensePlate->getLicensePlate());
                 if($blocker)
                 {
-                    $blockerEntry = $licensePlateRepository->findOneBy(['license_plate' => $blocker]);
-                    $mailer->sendBlockeeEmail($blockerEntry->getUser(), $entry->getUser(), $blockerEntry->getLicensePlate());
+                    foreach ($blocker as &$it)
+                    {
+                        $blockerEntry = $licensePlateRepository->findOneBy(['license_plate' => $it->getBlocker()]);
 
-                    $message = "Your car has been blocked by ".$blockerEntry->getLicensePlate()."!";
-                    $this->addFlash(
-                        'warning',
-                        $message
-                    );
+                        $mailer->sendBlockeeEmail($blockerEntry->getUser(), $entry->getUser(), $blockerEntry->getLicensePlate());
+
+                        $this->addFlash(
+                            'warning',
+                            "Your car has been blocked by ".$blockerEntry->getLicensePlate()."!"
+                        );
+
+                        $it->setStatus(1);
+                        $entityManager->persist($it);
+                        $entityManager->flush();
+                    }
                 }
 
                 $blockee = $activity->iveBlockedSomebody($licensePlate->getLicensePlate());
                 if($blockee)
                 {
-                    $blockeeEntry = $licensePlateRepository->findOneBy(['license_plate' => $blockee]);
-                    $mailer->sendBlockerEmail($blockeeEntry->getUser(), $entry->getUser(), $blockeeEntry->getLicensePlate());
-                    $this->addFlash('notice', 'Your were blocked by a car!');
+                    foreach ($blockee as &$it)
+                    {
+                        $blockeeEntry = $licensePlateRepository->findOneBy(['license_plate' => $it->getBlockee()]);
 
-                    $message="You blocked the car ".$blockeeEntry->getLicensePlate()."!";
-                    $this->addFlash(
-                        'danger',
-                        $message
-                    );
+                        $mailer->sendBlockerEmail($blockeeEntry->getUser(), $entry->getUser(), $blockeeEntry->getLicensePlate());
+
+                        $this->addFlash(
+                            'danger',
+                            "You blocked the car ".$blockeeEntry->getLicensePlate()."!"
+                        );
+
+                        $it->setStatus(1);
+                        $entityManager->persist($it);
+                        $entityManager->flush();
+                    }
                 }
 
                 return $this->redirectToRoute('license_plate_index');
@@ -108,19 +121,45 @@ class LicensePlateController extends AbstractController
     }
 
     #[Route('/{id}/edit', name: 'license_plate_edit', methods: ['GET', 'POST'])]
-    public function edit(Request $request, LicensePlate $licensePlate): Response
+    public function edit(Request $request, LicensePlate $licensePlate, LicensePlateService $licensePlateService, ActivityService $activityService): Response
     {
+        $oldLicensePlate = $licensePlate->getLicensePlate();
+
         $message = "Car ".$licensePlate->getLicensePlate()." has been changed to ";
 
         $form = $this->createForm(LicensePlateType::class, $licensePlate);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $licensePlate->setLicensePlate((new UnicodeString($licensePlate->getLicensePlate()))->camel()->upper());
-            $message = $message . $licensePlate->getLicensePlate();
+            $newLicensePlate = $licensePlateService->normalizeLicensePlate($licensePlate);
+
+            if($oldLicensePlate == $newLicensePlate)
+            {
+                $this->addFlash(
+                    'warning',
+                    'This license plate already exists!'
+                );
+
+                return $this->redirectToRoute('license_plate_index');
+            }
+
+            $blocker = $activityService->iveBlockedSomebody($oldLicensePlate);
+            $blockee = $activityService->whoBlockedMe($oldLicensePlate);
+
+            if($blocker != null || $blockee != null)
+            {
+                $this->addFlash(
+                    'warning',
+                    "Your license plate can't be changed. It already exists in a report."
+                );
+                return $this->redirectToRoute('license_plate_index');
+            }
+
+            $licensePlate->setLicensePlate($newLicensePlate);
+
             $this->addFlash(
                 'success',
-                $message
+                $message.' '.$licensePlate.'!'
             );
 
             $this->getDoctrine()->getManager()->flush();
@@ -135,12 +174,33 @@ class LicensePlateController extends AbstractController
     }
 
     #[Route('/{id}', name: 'license_plate_delete', methods: ['POST'])]
-    public function delete(Request $request, LicensePlate $licensePlate): Response
+    public function delete(Request $request, LicensePlate $licensePlate, ActivityService $activityService): Response
     {
+        $oldLicensePlate = $licensePlate->getLicensePlate();
+
+        $blocker = $activityService->iveBlockedSomebody($oldLicensePlate);
+        $blockee = $activityService->whoBlockedMe($oldLicensePlate);
+
+        if($blocker != null || $blockee != null)
+        {
+            $this->addFlash(
+                'warning',
+                "You can't delete your license plate! A report exists on your number."
+            );
+
+            return $this->redirectToRoute('license_plate_index');
+        }
+
         if ($this->isCsrfTokenValid('delete'.$licensePlate->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($licensePlate);
             $entityManager->flush();
+
+            $message = 'License plate ' . $licensePlate->getLicensePlate() . ' was successfully deleted!';
+            $this->addFlash(
+                'success',
+                $message
+            );
         }
 
         return $this->redirectToRoute('license_plate_index');
